scripts/requirements.txt
langchain==0.0.300
pinecone_client==2.2.1
python-dotenv==1.0.0